retrieval: cache dropped series, relable in place
Adds a cache of series that were dropped via relabeling to the scrape
cache so we don't need to recompute it on every scrape.
Relabeling and adding of target labels is now done within the scrape
rather than wrapped around the appender. This way we don't hold two
label representation in memory and can later move towards constructing
them in one go with the parsing.